test(architecture): move security boundary root tests
